Start on type detection.
