+++ tests/__fixtures__/examples/helpers.ts
@@ -1,2 +1,3 @@
 export function foo() {}
 export function bar() {}
+export type Baz = string;
